third_party/WebKit/LayoutTests/platform/android/external/wpt/web-animations/timing-model/animations/set-the-animation-start-time-expected.txt

This is a testharness.js-based test.
FAIL Setting the start time of an animation without an active timeline assert_equals: Setting the start time clears the current time expected (object) null but got (number) -463.4576666666703
FAIL Setting an unresolved start time an animation without an active timeline does not clear the current time assert_equals: Start time remains unresolved expected (object) null but got (number) 0
PASS Setting the start time clears the hold time
FAIL Setting an unresolved start time sets the hold time assert_approx_equals: Hold time is set after start time is made unresolved expected 1000 +/- 0.0005 but got 536.5423333333297
FAIL Setting the start time resolves a pending ready promise assert_true: Animation is in play-pending state expected true got undefined
FAIL Setting the start time resolves a pending pause task assert_true: Animation is in pause-pending state expected true got undefined
FAIL Setting the start time updates the finished state assert_greater_than: Setting the start time updated the finished state with the 'did seek' flag set to true expected a number greater than 100000 but got 100000
Harness: the test ran to completion.
